Update (ALPHA).
First commit containing working code. It's barely usable, and it kind of
works.. But it's definitely not yet finished; A crapload more work still
needs to be done. You can use it at this point if you want, but I'd
strongly advise against doing so in a production environment (better to
wait until a future commit containing more stable code is available),
and you should make sure that only you have access to your copy. I'll
write up some installation and usage instructions in the future, but I
haven't had time to do so yet (anyone with any PHP experience should be
able to work it out reasonably easily, though).